+++ SpatialGDK/Source/SpatialGDK/Public/SpatialGDKSettings.h
@@ -198,7 +198,7 @@ private:
 
 public:
 
-	bool GetPreventClientCloudDeploymentAutoConnect(bool bIsClient) const;
+	bool GetPreventClientCloudDeploymentAutoConnect() const;
 
 	UPROPERTY(EditAnywhere, Config, Category = "Region settings", meta = (ConfigRestartRequired = true, DisplayName = "Region where services are located"))
 	TEnumAsByte<EServicesRegion::Type> ServicesRegion;
@@ -210,10 +210,6 @@ public:
 	UPROPERTY(EditAnywhere, config, Category = "Debug", meta = (MetaClass = "SpatialDebugger"))
 	TSubclassOf<ASpatialDebugger> SpatialDebugger;
 
-	/** EXPERIMENTAL: Disable runtime load balancing and use a worker to do it instead. */
-	UPROPERTY(EditAnywhere, Config, Category = "Multi-Worker")
-	bool bEnableMultiWorker;
-
 	/** EXPERIMENTAL: Run SpatialWorkerConnection on Game Thread. */
 	UPROPERTY(Config)
 	bool bRunSpatialWorkerConnectionOnGameThread;
@@ -224,6 +220,8 @@ public:
 private:
 #if WITH_EDITOR
 	bool CanEditChange(const UProperty* InProperty) const override;
+
+	void UpdateServicesRegionFile();
 #endif
 
 	UPROPERTY(EditAnywhere, Config, Category = "Replication", meta = (DisplayName = "Use RPC Ring Buffers"))
@@ -273,6 +271,8 @@ public:
 
 	FORCEINLINE bool IsRunningInChina() const { return ServicesRegion == EServicesRegion::CN; }
 
+	void SetServicesRegion(EServicesRegion::Type NewRegion);
+
 	/** Enable to use the new net cull distance component tagging form of interest */
 	UPROPERTY(EditAnywhere, Config, Category = "Interest")
 	bool bEnableNetCullDistanceInterest;
@@ -289,12 +289,12 @@ public:
 	UPROPERTY(EditAnywhere, Config, Category = "Interest", meta = (EditCondition = "bEnableNetCullDistanceFrequency"))
 	TArray<FDistanceFrequencyPair> InterestRangeFrequencyPairs;
 
-	/** Use TLS encryption for UnrealClient workers connection. May impact performance. */
-	UPROPERTY(EditAnywhere, Config, Category = "Connection")
+	/** Use TLS encryption for UnrealClient workers connection. May impact performance. Only works in non-editor builds. */
+	UPROPERTY(EditAnywhere, Config, Category = "Connection", meta = (DisplayName = "Use Secure Client Connection In Packaged Builds"))
 	bool bUseSecureClientConnection;
 
-	/** Use TLS encryption for UnrealWorker (server) workers connection. May impact performance. */
-	UPROPERTY(EditAnywhere, Config, Category = "Connection")
+	/** Use TLS encryption for UnrealWorker (server) workers connection. May impact performance. Only works in non-editor builds. */
+	UPROPERTY(EditAnywhere, Config, Category = "Connection", meta = (DisplayName = "Use Secure Server Connection In Packaged Builds"))
 	bool bUseSecureServerConnection;
 
 	/**
@@ -308,4 +308,10 @@ public:
 	/** Experimental feature to use SpatialView layer when communicating with the Worker */
 	UPROPERTY(Config)
 	bool bUseSpatialView;
+
+	/**
+	  * By default, load balancing config will be read from the WorldSettings, but this can be toggled to override
+	  * the map's config with a 1x1 grid.
+	  */
+	TOptional<bool> bOverrideMultiWorker;
 };